Added better logging messaging, primarily debugs for better tracking execution

+++ caapkgmanager.cs
@@ -100,51 +100,63 @@ namespace aaPkgManager
         /// <param name="DestinationDirectory"></param>
         public void UnpackAAPKG(string AAPKGFilePath, string DestinationDirectory)
         {
+            log.Debug("AAPKGFilePath: " + AAPKGFilePath);
+            log.Debug("DestinationDirectory: " + DestinationDirectory);
+            log.Debug("begin");
 
             string file1Path;
             string outerUnpackPath;
 
             try
-            {
-                log.Debug("");
+            {             
 
                 // Calculate our outer unpack path
                 outerUnpackPath = Path.GetTempPath();
+                log.Debug("outerUnpackPath: " + outerUnpackPath);
 
                 // Go ahead and setup file1 path using the temp folder
                 file1Path = outerUnpackPath + "file1.cab";
+                log.Debug("file1Path: " + file1Path);
 
+                log.Debug("Outer extraction");
                 try
                 {
                     // Remove the outer AAPKG wrapper to reveal the inner file1.cab
                     using (ZipFile zip1 = ZipFile.Read(AAPKGFilePath))
                     {
+                        log.Debug("ExtractAll to " + outerUnpackPath);
                         zip1.ExtractAll(outerUnpackPath, ExtractExistingFileAction.OverwriteSilently);
                     }
                 }
                 catch (Ionic.Zip.ZipException)
                 {
+                    log.Debug("Exception using ExtractAll");
+                    log.Debug("Trying to use CabInfo.UnPack");
                     // If we have an error reading as a zip then try to unpack as a CAB
                     CabInfo ci = new CabInfo(AAPKGFilePath);
                     ci.Unpack(outerUnpackPath);
                 }
 
-
+                log.Debug("inner extraction");
                 // Extract the innner file1.cab to the destination directory
                 try
                 {
                     using (ZipFile zip2 = ZipFile.Read(file1Path))
                     {
+                        log.Debug("ExtractAll to " + DestinationDirectory);
                         zip2.ExtractAll(DestinationDirectory, ExtractExistingFileAction.OverwriteSilently);
                     }
                 }
                 catch (Ionic.Zip.ZipException)
                 {
+                    log.Debug("Exception using ExtractAll");
+                    log.Debug("Trying to use CabInfo.UnPack for " + file1Path);
                     // If we have an error reading as a zip then try to unpack as a CAB
                     CabInfo ci = new CabInfo(file1Path);
                     ci.Unpack(DestinationDirectory);
                 }
 
+                log.Debug("Delete path " + file1Path);
                 // Delete the file1.cab
                 System.IO.File.Delete(file1Path);
 
@@ -162,19 +174,20 @@ namespace aaPkgManager
         /// <returns></returns>
         public string UnpackAAPKG(string AAPKGFilePath)
         {
+            log.Debug("AAPKGFilePath: " + AAPKGFilePath);
+            log.Debug("begin");
+
             string tempPath;
 
             try
             {
-
-                log.Debug("");
-
                 // First extract the objects to a temporary path
                 tempPath = Path.GetTempPath() + System.Guid.NewGuid();
+                log.Debug("tempPath: " + tempPath);
 
                 // Delete the folder if it already exists.. unlikely since we are using a GUID but just to be safe
-                if (System.IO.Directory.Exists(tempPath)) { System.IO.Directory.Delete(tempPath, true); }
-
+                if (System.IO.Directory.Exists(tempPath)) { System.IO.Directory.Delete(tempPath, true); log.Debug("Deleted " + tempPath); }
+                
                 // Perform the unpack
                 this.UnpackAAPKG(AAPKGFilePath, tempPath);
 
@@ -187,28 +200,80 @@ namespace aaPkgManager
             }
 
         }
+        
+        public MemoryStream UnPackStreamAsStream(MemoryStream inputStream, string fileName)
+        {
+            try
+            {
+                MemoryStream resultStream = new MemoryStream();
+                using (ZipFile zip = ZipFile.Read(inputStream))
+                {
+                    zip[fileName].Extract(resultStream);
+                }
+                resultStream.Seek(0, SeekOrigin.Begin);
+                return resultStream;
+            }
+            catch
+            {
+                throw;
+            }
+        }
+
+        public MemoryStream UnpackFileAsStream(string filePath, string fileName)
+        {            
+            try
+            {
+                MemoryStream resultStream = new MemoryStream();
+                using (ZipFile zip = ZipFile.Read(filePath))
+                {
+                    zip[fileName].Extract(resultStream);
+                }
+                resultStream.Seek(0, SeekOrigin.Begin);
+                return resultStream;
+            }
+            catch
+            {
+                throw;
+            }
+
+
+        }
 
         /// <summary>
         /// Get the manifest as a string
         /// </summary>
         /// <param name="AAPKGFilePath"></param>
+        /// <param name="InMemory"></param>
         /// <returns></returns>
         /// TODO: Use in memory instead of extracting to disk then cleaning up
-        public string GetManifestAsString(string AAPKGFilePath)
+        public string GetManifestAsString(string AAPKGFilePath,bool InMemory= true)
         {
-            string workingPath;
             string manifestString;
 
             try
             {
-                // Unpack the files
-                workingPath = this.UnpackAAPKG(AAPKGFilePath);
 
-                // Get the XML from the manifest file
-                manifestString = File.ReadAllText(workingPath + "\\manifest.xml");
+                if (InMemory)
+                {
+                    MemoryStream File1ms;
+                    MemoryStream Manifestms;
 
-                //Delete the folder
-                Directory.Delete(workingPath, true);
+                    File1ms = this.UnpackFileAsStream(AAPKGFilePath, "file1.cab");
+                    Manifestms = this.UnPackStreamAsStream(File1ms, "manifest.xml");
+
+                    manifestString = new StreamReader(Manifestms).ReadToEnd();
+                }
+                else
+                {                    
+                    // Unpack the files
+                    string workingPath = this.UnpackAAPKG(AAPKGFilePath);
+
+                    // Get the XML from the manifest file
+                    manifestString = File.ReadAllText(workingPath + "\\manifest.xml");
+
+                    //Delete the folder
+                    Directory.Delete(workingPath, true);
+                }
 
                 //return the result
                 return manifestString;
@@ -262,35 +327,50 @@ namespace aaPkgManager
 
                 // Extract to a temporary location
                 workingPath = this.UnpackAAPKG(AAPKGFilePath);
+                log.Debug("workingPath: " + workingPath);
 
                 //Calculate the path to the manifest
                 manifestFileName = workingPath + @"\Manifest.xml";
+                log.Debug("manifestFileName: " + manifestFileName);
 
                 // Load the Manifest File
                 XDocument xmanifestDoc = XDocument.Load(manifestFileName);
+                log.Debug("xmanifestDoc: ");
+                log.Debug(xmanifestDoc.ToString());
+
 
                 // If objects is null then create a blank
                 if (ObjectsToKeep == null)
                 {
                     ObjectsToKeep = new List<String>();
+                    log.Debug("ObjectsToKeep Null.  Set to new List<String>()");
                 }
+
+
+                log.Debug("ObjectsToKeep: " + ObjectsToKeep.ToString());
 
                 // If the Objects to Keep list is not null then remove items based on the list
                 if (ObjectsToKeep.Count() > 0)
-                {
+                {                    
                     //Remove items from aaPKG file based on the objects to keep list that has been passed
                     this.RemoveItemsFromUnpackedFiles(workingPath, ref xmanifestDoc, ObjectsToKeep);
                 }
 
+                log.Debug("ExcludeAAPDFFiles: " + ExcludeAAPDFFiles);
                 // Remove the aaPDF files if required
-                if (ExcludeAAPDFFiles) { this.RemoveAAPDFFilesFromUnpackedFiles(workingPath, ref xmanifestDoc); }
+                if (ExcludeAAPDFFiles) { this.RemoveAAPDFFilesFromUnpackedFiles(workingPath, ref xmanifestDoc);}
 
                 // Save the doc back
                 xmanifestDoc.Save(manifestFileName);
+                log.Debug("Saved xmanifestDoc");
+                log.Debug("xmanifestDoc: ");
+                log.Debug(xmanifestDoc.ToString());
+
 
                 //If we don't have a new aapkg name then just reuse the old name
                 if (NewAAPKGFileName == "")
                 {
+                    log.Debug("NewAAPKGFileName blank. Setting to " + AAPKGFilePath);
                     NewAAPKGFileName = AAPKGFilePath;
                 }
 
@@ -299,6 +379,7 @@ namespace aaPkgManager
 
                 //Delete the working path
                 if (System.IO.Directory.Exists(workingPath)) { System.IO.Directory.Delete(workingPath, true); }
+                log.Debug("workingPath " + workingPath + " deleted");
             }
             catch
             {
@@ -429,37 +510,47 @@ namespace aaPkgManager
         /// <param name="xmanifestDoc"></param>
         private void RemoveAAPDFFilesFromUnpackedFiles(string workingPath, ref XDocument xmanifestDoc)
         {
+            log.Debug("workingPath: " + workingPath);
+            log.Debug("xmanifestDoc: ");
+            log.Debug(xmanifestDoc.ToString());
+            
             string dirToDelete;
 
             try
             {
-                log.Debug("");
-
+            
                 // Loop through all of the AAPDF files in the folder and create a 0 size file in it's place.
                 // We have to do this b/c the import will fail if the file is missing
                 // However, if the file is present but the importer never reads it then the file passes
                 string[] aaPDFFiles;
 
                 aaPDFFiles = Directory.GetFiles(workingPath, "*.aapdf");
+                log.Debug("aaPDFFiles: " + aaPDFFiles.ToString());
 
                 //Loop through the filenames and create a blank file, overwriting the aaPDF in place
                 foreach (string filename in aaPDFFiles)
                 {
+                    log.Debug("filename: " + filename);
+
                     // Get the XML node that references this aaPDF File
                     var pdfElement = xmanifestDoc.XPathSelectElement("//template[@file_name=\"" + Path.GetFileName(filename) + "\"]");
+                    log.Debug("pdfElement: " + pdfElement.ToString());
 
                     // Crate the blank file in place of the aaPDF File
                     File.Create(filename).Close();
+                    log.Debug("Created zero file " + filename);
 
                     //Clear the .txt file
                     File.Create(workingPath + "\\" + pdfElement.Attribute("tag_name").Value + ".txt").Close();
-
+                    log.Debug("Created zero file " + workingPath + "\\" + pdfElement.Attribute("tag_name").Value + ".txt");
+                    
                     // Set the version to -1 to make sure it doesn't get imported
                     pdfElement.SetAttributeValue("config_version", "-1");
+                    log.Debug("set config_version to -1");
 
                     //Delete the help folder
                     dirToDelete = workingPath + "\\" + pdfElement.Attribute("gobjectid").Value;
-                    if (System.IO.Directory.Exists(dirToDelete)) { System.IO.Directory.Delete(dirToDelete, true); }
+                    if (System.IO.Directory.Exists(dirToDelete)) { System.IO.Directory.Delete(dirToDelete, true); log.Debug("Deleted " + dirToDelete); }
 
                 }
 
@@ -478,10 +569,14 @@ namespace aaPkgManager
         /// <param name="ObjectsToKeep"></param>
         private void RemoveItemsFromUnpackedFiles(string workingPath, ref XDocument xmanifestDoc, List<String> ObjectsToKeep)
         {
+            log.Debug("workingPath: " + workingPath);
+            log.Debug("xmanifestDoc: ");
+            log.Debug(xmanifestDoc.ToString());
+            log.Debug("ObjectsToKeep: " + ObjectsToKeep.ToString());
+            log.Debug("begin");
+
             try
             {
-                log.Debug("");
-
                 //Use a compound query for templates and instances
                 this.RemoveItemsFromUnpackedFilesByXPath(workingPath, ref xmanifestDoc, ObjectsToKeep, "//derived_templates/template | //derived_instances/instance");
             }
@@ -544,10 +639,17 @@ namespace aaPkgManager
         {
             try
             {
-                log.Debug("");
+                log.Debug("workingPath: " + workingPath);
+                log.Debug("xmanifestDoc: ");
+                log.Debug(xmanifestDoc.ToString());
+                log.Debug("ObjectsToKeep: " + ObjectsToKeep.ToString());
+                log.Debug("XPathQuery: " + XPathQuery);
+                log.Debug("begin");
 
                 //Templates
                 var elements = xmanifestDoc.XPathSelectElements(XPathQuery);
+                log.Debug("elements");
+                log.Debug(elements.ToString());
 
                 foreach (XElement element in elements)
                 {
@@ -556,12 +658,14 @@ namespace aaPkgManager
                     {
                         if (ObjectsToKeep.IndexOf(element.Attributes("tag_name").Single().Value) == -1)
                         {
+                            log.Debug("removing " + element.Attributes("tag_name"));
                             element.SetAttributeValue("config_version", "-1");
                             this.DeleteGObjectFiles(workingPath, element.Attribute("gobjectid").Value);
                         }
                     }
                     else
                     {
+                        log.Debug("removing " + element.Attributes("tag_name"));
                         element.SetAttributeValue("config_version", "-1");
                         this.DeleteGObjectFiles(workingPath, element.Attribute("gobjectid").Value);
                     }
@@ -580,16 +684,20 @@ namespace aaPkgManager
         /// <param name="gObjectID"></param>
         private void DeleteGObjectFiles(string workingPath, string gObjectID)
         {
-            try
-            {
-                log.Debug("");
+            log.Debug("workingPath: " + workingPath);
+            log.Debug("gObjectID: " + gObjectID);
+            log.Debug("begin");
 
+            try
+            {                
                 // Zero out the file and immediately close
                 File.Create(workingPath + "\\" + gObjectID + ".txt").Close();
+                log.Debug("Created zero length file at " + workingPath + "\\" + gObjectID + ".txt");
 
                 // Delete dir if it exists
                 string dirToDelete = workingPath + "\\" + gObjectID;
-                if (System.IO.Directory.Exists(dirToDelete)) { System.IO.Directory.Delete(dirToDelete, true); }
+                log.Debug("Attempting to delete " + dirToDelete);
+                if (System.IO.Directory.Exists(dirToDelete)) { System.IO.Directory.Delete(dirToDelete, true); log.Debug("deleted " + dirToDelete);}
             }
             catch
             {
